+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Fix bug where `template/config.yaml` extension value isn't checked
   properly when the extension has a dot: eg `extension = .theme.json`
 - README.md links
+- Builder now reads scheme slug instead of always inferring based on
+  scheme name
 
 ### Updated